Rename CEO -> Manager
* Move roles to their own file and remove common enum file
