+++ alcs-api/src/application/application-code/application-code.controller.ts
@@ -4,8 +4,8 @@ import { Controller, Get, UseGuards } from '@nestjs/common';
 import { ApiOAuth2 } from '@nestjs/swagger';
 import * as config from 'config';
 import { RoleGuard } from '../../common/authorization/role.guard';
+import { ANY_AUTH_ROLE } from '../../common/authorization/roles';
 import { UserRoles } from '../../common/authorization/roles.decorator';
-import { ANY_AUTH_ROLE } from '../../common/enum';
 import { ApplicationStatusDto } from '../application-status/application-status.dto';
 import { ApplicationStatus } from '../application-status/application-status.entity';
 import { ApplicationMasterCodesDto } from './application-code.dto';
