+++ alcs-api/src/application/application.controller.ts
@@ -13,12 +13,11 @@ import {
 import { ApiOAuth2 } from '@nestjs/swagger';
 import * as config from 'config';
 import { RoleGuard } from '../common/authorization/role.guard';
+import { ANY_AUTH_ROLE } from '../common/authorization/roles';
 import { UserRoles } from '../common/authorization/roles.decorator';
-import { ANY_AUTH_ROLE } from '../common/enum';
 import { ServiceValidationException } from '../common/exceptions/base.exception';
 import { ApplicationCodeService } from './application-code/application-code.service';
 import { ApplicationDecisionMaker } from './application-code/application-decision-maker/application-decision-maker.entity';
-import { ApplicationRegionDto } from './application-code/application-region/application-region.dto';
 import { ApplicationRegion } from './application-code/application-region/application-region.entity';
 import { ApplicationType } from './application-code/application-type/application-type.entity';
 import { ApplicationStatus } from './application-status/application-status.entity';
