+++ alcs-api/src/common/authorization/roles.decorator.ts
@@ -1,5 +1,5 @@
 import { SetMetadata } from '@nestjs/common';
-import { AUTH_ROLE } from '../enum';
+import { AUTH_ROLE } from './roles';
 
 export const UserRoles = (...roles: AUTH_ROLE[]) =>
   SetMetadata('userRoles', roles);
