+++ alcs-api/src/user/user.controller.ts
@@ -4,8 +4,8 @@ import { Body, Controller, Delete, Get, UseGuards } from '@nestjs/common';
 import { ApiOAuth2 } from '@nestjs/swagger';
 import * as config from 'config';
 import { RoleGuard } from '../common/authorization/role.guard';
+import { ANY_AUTH_ROLE, AUTH_ROLE } from '../common/authorization/roles';
 import { UserRoles } from '../common/authorization/roles.decorator';
-import { ANY_AUTH_ROLE, AUTH_ROLE } from '../common/enum';
 import { UserDto } from './user.dto';
 import { User } from './user.entity';
 import { UserService } from './user.service';
